+++ src/common/component/DatePicker/index.tsx
@@ -1,7 +1,4 @@
-import { isBefore } from 'date-fns';
-
-import { useState } from 'react';
-
+import { useDatePicker } from '@/common/hook/useDatePicker';
 import { useOutsideClick } from '@/common/hook/useOutsideClick';
 import { useOverlay } from '@/common/hook/useOverlay';
 
@@ -15,28 +12,7 @@ interface DatePickerProps {
 const DatePicker = ({ variant }: DatePickerProps) => {
   const { isOpen, close, toggle } = useOverlay();
   const ref = useOutsideClick<HTMLDivElement>(close);
-
-  const [selectedDate, setSelectedDate] = useState<Date | null>(null);
-  const [endDate, setEndDate] = useState<Date | null>(null);
-
-  const handleSelectDate = (date: Date) => {
-    if (variant === 'range') {
-      // 두 날짜가 선택되었을 때, 자동으로 순서를 조정 (시작 날짜가 종료 날짜보다 늦으면 교환)
-      if (!selectedDate || (selectedDate && endDate)) {
-        setSelectedDate(date);
-        setEndDate(null);
-      } else {
-        if (isBefore(date, selectedDate)) {
-          setEndDate(selectedDate);
-          setSelectedDate(date);
-        } else {
-          setEndDate(date);
-        }
-      }
-    } else {
-      setSelectedDate(date); // 단일 날짜 설정
-    }
-  };
+  const { selectedDate, endDate, handleSelectDate } = useDatePicker(variant);
 
   return (
     <div ref={ref} style={{ position: 'relative', display: 'flex', flexDirection: 'column', gap: '0.4rem' }}>
@@ -47,14 +23,12 @@ const DatePicker = ({ variant }: DatePickerProps) => {
         variant={variant}
         width="10.3rem"
       />
-
       {isOpen && (
         <DatePickerCalendar
           selectedDate={selectedDate || new Date()}
           endDate={endDate}
           setSelectedDate={handleSelectDate}
           variant={variant}
-          isOpen={isOpen}
         />
       )}
     </div>
